Arrange action refactored
I renamed the BringToFront action to SendToFront to keep the arrange actions names uniform. In the nor SendToFront and SendToBack action classes I extracted a method in both classes which also left the original method as composed method where the level of details were uniform throughout the method.

I have tried to change the label property files so my change of the BringToFront name also can be seen in the GUI as i want to keep it consistent in the whole system. I think i have located the specific file where the change is supposed to go, but I am not seeing the change in the GUI - for now I don't know why. Maybe some cache stuff?